docs: fix typos (#26)
* fix typo getting-started.mdx

- "isntance" -> "instance"

"isntance" is a typographical error. The correct spelling is "instance."

* fix typo gas-fees.mdx

"their" -> "there"

"their" is a possessive pronoun, while "there" is used to indicate a place or position. The correct usage in this context is "there".

* fix typo evm-opcodes.mdx

"Absrtact" -> "Abstract"

"Absrtact" is a typographical error. The correct spelling is "Abstract."

* fix typo using-system-contracts.mdx

- "Undersand" -> "Understand"

"Undersand" is a typographical error. The correct spelling is "Understand."

* fix typo list-of-system-contracts.mdx

- "Why am i getting "the code hash is not known" error?" -> "Why am I getting "the code hash is not known" error?"

"I" should be capitalized to follow standard English grammar rules.

* fix typo handling-nonces.mdx

- "tranasction" -> "transaction"

"tranasction" is a typographical error. The correct spelling is "transaction."